refactor RootState interface #GSM-62
