+++ frontend/components/header/priceOutputContainer.tsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import PriceOutput from './priceOutput';
-import { RootState } from '../../utilities/interfaces';
+import { RootState } from '../../reducers/rootReducer';
 
 const mapStateToProps = (state: RootState) => ({
   latestPrice: state.quotes.companyStats.latestPrice,
