+++ frontend/components/header/searchInputContainer.tsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import SearchInput from './searchInput';
 import { searchAction, fetchCompanyNames } from '../../actions/quotesActions';
-import { RootState } from '../../utilities/interfaces';
+import { RootState } from '../../reducers/rootReducer';
 import { ThunkDispatch } from 'redux-thunk';
 
 const mapStateToProps = ({ quotes }: RootState) => ({
