+++ frontend/components/statsInfoPeers/keyStatsListContainer.tsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import KeyStatsList from './keyStatsList';
-import { RootState } from '../../utilities/interfaces';
+import { RootState } from '../../reducers/rootReducer';
 import {
   selectCompanyStats,
   selectFetchingStatus,
