+++ frontend/utilities/selectors.ts
@@ -1,9 +1,5 @@
-import {
-  CompanyStatsState,
-  ChartDataDay,
-  ChartData,
-  RootState,
-} from './interfaces';
+import { CompanyStatsState, ChartDataDay, ChartData } from './interfaces';
+import { RootState } from '../reducers/rootReducer';
 import { changeToPercent } from './numberFormatters';
 
 export const selectCompanyStats = (companyStats: CompanyStatsState) => {
